Move step to Linux
