+++ src/main.rs
@@ -251,6 +251,9 @@ fn run() -> Result<()> {
         runner.execute(Step::Lure, "LURE", || linux::run_lure_update(&ctx))?;
         runner.execute(Step::Waydroid, "Waydroid", || linux::run_waydroid(&ctx))?;
         runner.execute(Step::AutoCpufreq, "auto-cpufreq", || linux::run_auto_cpufreq(&ctx))?;
+        runner.execute(Step::CinnamonSpices, "Cinnamon spices", || {
+            linux::run_cinnamon_spices_updater(&ctx)
+        })?;
     }
 
     #[cfg(target_os = "macos")]
@@ -442,9 +445,6 @@ fn run() -> Result<()> {
     runner.execute(Step::Aqua, "aqua", || generic::run_aqua(&ctx))?;
     runner.execute(Step::Bun, "bun", || generic::run_bun(&ctx))?;
     runner.execute(Step::Zigup, "zigup", || generic::run_zigup(&ctx))?;
-    runner.execute(Step::CinnamonSpices, "Cinnamon spices", || {
-        generic::run_cinnamon_spices_updater(&ctx)
-    })?;
 
     if should_run_powershell {
         runner.execute(Step::Powershell, "Powershell Modules Update", || {
