+++ src/steps/generic.rs
@@ -1303,14 +1303,3 @@ pub fn run_zigup(ctx: &ExecutionContext) -> Result<()> {
 
     Ok(())
 }
-
-pub fn run_cinnamon_spices_updater(ctx: &ExecutionContext) -> Result<()> {
-    let cinnamon_spice_updater = require("cinnamon-spice-updater")?;
-
-    print_separator("Cinnamon spices");
-
-    ctx.run_type()
-        .execute(cinnamon_spice_updater)
-        .arg("--update-all")
-        .status_checked()
-}
